+++ system/include/emscripten/webaudio.h
@@ -19,10 +19,16 @@ extern "C" {
 
 typedef int EMSCRIPTEN_WEBAUDIO_T;
 
+// Default render size of 128 frames
+#define AUDIO_CONTEXT_RENDER_SIZE_DEFAULT 0
+// Let the hardware determine the best render size
+#define AUDIO_CONTEXT_RENDER_SIZE_HARDWARE -1
+
 typedef struct EmscriptenWebAudioCreateAttributes
 {
 	const char *latencyHint; // Specify one of "balanced", "interactive" or "playback"
 	uint32_t sampleRate; // E.g. 44100 or 48000
+	int32_t renderSizeHint; // AUDIO_CONTEXT_RENDER_SIZE_* or number of samples
 } EmscriptenWebAudioCreateAttributes;
 
 // Creates a new Web Audio AudioContext, and returns a handle to it.
@@ -96,7 +102,7 @@ typedef void (*EmscriptenWorkletProcessorCreatedCallback)(EMSCRIPTEN_WEBAUDIO_T 
 void emscripten_create_wasm_audio_worklet_processor_async(EMSCRIPTEN_WEBAUDIO_T audioContext, const WebAudioWorkletProcessorCreateOptions *options, EmscriptenWorkletProcessorCreatedCallback callback, void *userData3);
 
 // Returns the number of samples processed per channel in an AudioSampleFrame, fixed at 128 in the Web Audio API 1.0 specification, and valid for the lifetime of the audio context.
-// For this to change from the default 128, the context would need to be created with a yet unexposed WebAudioWorkletProcessorCreateOptions renderSizeHint, part of the 1.1 Web Audio API.
+// For this to differ from the default 128, the context would need to be created with a WebAudioWorkletProcessorCreateOptions renderSizeHint, part of the 1.1 Web Audio API.
 int emscripten_audio_context_quantum_size(EMSCRIPTEN_WEBAUDIO_T audioContext);
 
 // Returns the sampling rate of the given Audio Context, e.g. 48000 or 44100 or similar.