+++ test/runner.py
@@ -25,7 +25,6 @@ import logging
 import math
 import operator
 import os
-import platform
 import random
 import sys
 import time
@@ -41,9 +40,11 @@ import browser_common
 import common
 import jsrun
 import parallel_testsuite
+from color_runner import ColorTextRunner
 from common import errlog
+from single_line_runner import SingleLineTestRunner
 
-from tools import config, shared, utils
+from tools import colored_logger, config, shared, utils
 
 logger = logging.getLogger("runner")
 
@@ -420,9 +421,7 @@ def run_tests(options, suites):
     print('Test suites:', [s[0] for s in suites])
   # Run the discovered tests
 
-  # We currently don't support xmlrunner on macOS M1 runner since
-  # `pip` doesn't seeem to yet have pre-built binaries for M1.
-  if os.getenv('CI') and not (utils.MACOS and platform.machine() == 'arm64'):
+  if os.getenv('CI'):
     os.makedirs('out', exist_ok=True)
     # output fd must remain open until after testRunner.run() below
     output = open('out/test-results.xml', 'wb')
@@ -430,8 +429,16 @@ def run_tests(options, suites):
     testRunner = xmlrunner.XMLTestRunner(output=output, verbosity=2,
                                          failfast=options.failfast)
     print('Writing XML test output to ' + os.path.abspath(output.name))
+  elif options.ansi and not options.verbose:
+    # When not in verbose mode and ansi color output is available use our nice single-line
+    # result display.
+    testRunner = SingleLineTestRunner(failfast=options.failfast)
   else:
-    testRunner = unittest.TextTestRunner(verbosity=2, failfast=options.failfast)
+    if not options.ansi:
+      print('using verbose test runner (ANSI not avilable)')
+    else:
+      print('using verbose test runner (verbose output requested)')
+    testRunner = ColorTextRunner(failfast=options.failfast)
 
   total_core_time = 0
   run_start_time = time.perf_counter()
@@ -469,7 +476,12 @@ def parse_args():
                            'test.  Implies --cores=1.  Defaults to true when running a single test')
   parser.add_argument('--no-clean', action='store_true',
                       help='Do not clean the temporary directory before each test run')
-  parser.add_argument('--verbose', '-v', action='store_true')
+  parser.add_argument('--verbose', '-v', action='count', default=0,
+                      help="Show test stdout and stderr, and don't use the single-line test reporting. "
+                           'Specifying `-v` twice will enable test framework logging (i.e. EMTEST_VERBOSE)')
+  # TODO: Replace with BooleanOptionalAction once we can depend on python3.9
+  parser.add_argument('--ansi', action='store_true', default=None)
+  parser.add_argument('--no-ansi', action='store_false', dest='ansi', default=None)
   parser.add_argument('--all-engines', action='store_true')
   parser.add_argument('--detect-leaks', action='store_true')
   parser.add_argument('--skip-slow', action='store_true', help='Skip tests marked as slow')
@@ -500,6 +512,14 @@ def parse_args():
   parser.add_argument('--bell', action='store_true', help='Play a sound after the test suite finishes.')
 
   options = parser.parse_args()
+
+  if options.ansi is None:
+    options.ansi = colored_logger.ansi_color_available()
+  else:
+    if options.ansi:
+      colored_logger.enable(force=True)
+    else:
+      colored_logger.disable()
 
   if options.failfast:
     if options.max_failures != 2**31 - 1:
@@ -572,7 +592,7 @@ def main():
   set_env('EMTEST_SKIP_SLOW', options.skip_slow)
   set_env('EMTEST_ALL_ENGINES', options.all_engines)
   set_env('EMTEST_REBASELINE', options.rebaseline)
-  set_env('EMTEST_VERBOSE', options.verbose)
+  set_env('EMTEST_VERBOSE', options.verbose > 1)
   set_env('EMTEST_CORES', options.cores)
   set_env('EMTEST_FORCE64', options.force64)
 
@@ -583,20 +603,8 @@ def main():
   # Remove any old test files before starting the run
   cleanup_emscripten_temp()
   utils.delete_file(common.flaky_tests_log_filename)
-  utils.delete_file(browser_common.browser_spawn_lock_filename)
-  utils.delete_file(f'{browser_common.browser_spawn_lock_filename}_counter')
-  if options.force_browser_process_termination or os.getenv('EMTEST_FORCE_BROWSER_PROCESS_TERMINATION'):
-    config = common.get_browser_config()
 
-    def terminate_all_browser_processes():
-      procs = common.list_processes_by_name(config.executable_name)
-      if len(procs) > 0:
-        print(f'Terminating {len(procs)} stray browser processes.')
-        common.terminate_list_of_processes(procs)
-
-    if config and hasattr(config, 'executable_name'):
-      atexit.register(terminate_all_browser_processes)
-      terminate_all_browser_processes()
+  browser_common.init(options.force_browser_process_termination)
 
   def prepend_default(arg):
     if arg.startswith('test_'):