+++ test/test_sanity.py
@@ -8,6 +8,7 @@ import os
 import re
 import shutil
 import stat
+import sys
 import time
 from pathlib import Path
 from subprocess import PIPE, STDOUT
@@ -22,7 +23,7 @@ from common import (
   path_from_root,
   test_file,
 )
-from decorators import crossplatform, parameterized, with_env_modify
+from decorators import crossplatform, no_windows, parameterized, with_env_modify
 
 from tools import cache, ports, response_file, shared, utils
 from tools.config import EM_CONFIG
@@ -237,6 +238,7 @@ class sanity(RunnerCore):
         finally:
           delete_file(default_config)
 
+  @no_windows('Test relies on Unix-specific make_fake_tool')
   def test_llvm(self):
     LLVM_WARNING = 'LLVM version for clang executable'
 
@@ -422,6 +424,7 @@ fi
     self.assertIn(SANITY_MESSAGE, output)
     self.assertCacheEmpty()
 
+  @no_windows('Test relies on Unix-specific make_fake_tool')
   def test_cache_clearing_auto(self):
     # Changing LLVM_ROOT, even without altering .emscripten, clears the cache
     restore_and_set_up()
@@ -639,6 +642,7 @@ fi
       build()
       test()
 
+  @no_windows('Test relies on Unix-specific make_fake_tool')
   def test_vanilla(self):
     restore_and_set_up()
     self.clear_cache()
@@ -663,6 +667,7 @@ fi
     test_with_fake('got js backend! JavaScript (asm.js, emscripten) backend', 'LLVM has not been built with the WebAssembly backend')
     delete_dir(shared.CANONICAL_TEMP_DIR)
 
+  @no_windows('Test relies on Unix-specific make_fake_tool')
   def test_llvm_add_version(self):
     restore_and_set_up()
 
@@ -689,6 +694,7 @@ fi
     open(EM_CONFIG, 'a').write('\ndel BINARYEN_ROOT\n')
     self.check_working([EMCC, test_file('hello_world.c')], 'BINARYEN_ROOT not set in config (%s), and `wasm-opt` not found in PATH' % EM_CONFIG)
 
+  @no_windows('Test relies on Unix-specific make_fake_tool')
   def test_empty_config(self):
     restore_and_set_up()
     make_fake_tool(self.in_dir('fake', 'wasm-opt'), 'foo')
@@ -699,6 +705,7 @@ fi
     with env_modify({'PATH': self.in_dir('fake') + os.pathsep + os.environ['PATH']}):
       self.check_working([EMCC])
 
+  @no_windows('Test relies on Unix-specific make_fake_tool')
   def test_missing_config(self):
     restore_and_set_up()
     make_fake_tool(self.in_dir('fake', 'wasm-opt'), 'foo')
@@ -786,6 +793,7 @@ fi
     self.run_process([EMBUILDER, '--pic', 'clear', 'sdl2*'])
     self.run_process([EMCC, '-sMAIN_MODULE=2', '-sUSE_SDL=2', '-sUSE_SDL_GFX=2', test_file('hello_world.c')])
 
+  @no_windows('Test relies on Unix-specific make_fake_tool')
   def test_binaryen_version(self):
     restore_and_set_up()
     with open(EM_CONFIG, 'a') as f:
@@ -822,9 +830,20 @@ fi
     for e in ['LLVM_ROOT', 'EMSDK_NODE', 'EMSDK_PYTHON', 'EMSDK', 'EMSCRIPTEN', 'BINARYEN_ROOT', 'EMCC_SKIP_SANITY_CHECK', 'EM_CONFIG']:
       env.pop(e, None)
 
+    python_path = os.path.normpath(os.path.dirname(sys.executable))
+
     # Remove from PATH every directory that contains clang.exe so that bootstrap.py cannot
     # accidentally succeed by virtue of locating tools in PATH.
-    new_path = [d for d in env['PATH'].split(os.pathsep) if not os.path.isfile(os.path.join(d, utils.exe_suffix('clang')))]
+    def ignore_path(p):
+      clang_bin = utils.exe_suffix('clang')
+      # We cannot ignore a path element that contains the python executable itself, otherwise
+      # the bootstrap script will fail
+      if os.path.isfile(os.path.join(p, clang_bin)) and os.path.normpath(p) != python_path:
+        return True
+      return False
+
+    old_path = env['PATH'].split(os.pathsep)
+    new_path = [d for d in old_path if not ignore_path(d)]
     env['PATH'] = os.pathsep.join(new_path)
 
     # Running bootstrap.py should not fail